pyproject.toml
[project]
name = "ragsystem"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "llama-index>=0.12.36",
    "llama-index-readers-wikipedia>=0.3.0",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
    "wikipedia>=1.4.0",
]
